requirements.txt
asgiref==3.5.0
bleach==5.0.0
boto3==1.22.4
botocore==1.25.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-storages==1.12.3
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
